backend/requirements.txt
# =============================================================================
# CodeForge AI — Python Backend Dependencies
# =============================================================================
# Install all at once:  pip install -r requirements.txt
# Python version:       3.11+
# =============================================================================

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0          # High-performance async web framework
uvicorn[standard]>=0.30.0  # ASGI server (run FastAPI)
python-multipart>=0.0.9   # Needed for form data parsing

# ── AI / LLM ──────────────────────────────────────────────────────────────────
langchain>=0.3.0           # LLM orchestration framework
langchain-openai>=0.2.0    # OpenAI-compatible LLM wrapper (used for OpenRouter)
langgraph>=0.2.0           # Multi-agent workflow graph (Directed Acyclic Graph)

# ── Database ──────────────────────────────────────────────────────────────────
motor>=3.5.0               # Async MongoDB driver (non-blocking)
pymongo>=4.8.0             # MongoDB driver (motor dependency)
qdrant-client>=1.11.0      # Qdrant vector store client

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0  # JWT token creation and verification
passlib[bcrypt]>=1.7.4             # Password hashing (bcrypt)

# ── Networking ────────────────────────────────────────────────────────────────
httpx>=0.27.0             # Async HTTP client (GitHub API, Twilio, etc.)
websockets>=13.0          # WebSocket support
sse-starlette>=2.1.0      # Server-Sent Events (SSE) for AI streaming

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic-settings>=2.5.0  # Settings management from environment variables
python-dotenv>=1.0.0      # Load .env file into environment

# ── Messaging / Integrations ──────────────────────────────────────────────────
twilio>=9.3.0             # WhatsApp + Instagram messaging via Twilio API
